Fix bug of removing and refactoring some parts

+++ app/src/main/java/com/novelitech/practicetodokotlin/dataclasses/Todo.kt
@@ -8,7 +8,7 @@ data class Todo (
     val done: Boolean,
 
     @Transient
-    val onRemove: () -> Unit,
+    val onRemove: (Int) -> Unit,
 
     @Transient
     val onCheck: (Boolean) -> Unit
